power: refactor component in order to dynamically import plugins just by adding them to the extras/power directory
