+++ moonraker/extras/__init__.py
@@ -1,0 +1,1 @@
+# package definition for thirdparty package
