+++ moonraker/extras/power/__init__.py
@@ -1,0 +1,7 @@
+# package definition for thirdparty package
+import glob
+from os.path import dirname, basename, isfile, join
+
+modules = glob.glob(join(dirname(__file__), "*.py"))
+__all__ = [basename(f)[:-3]
+           for f in modules if isfile(f) and not f.endswith('__init__.py')]
